[git] pack: fix binary search, add comments
We checked if 'midpoint' was the desired sha1, and it was not, but
smaller, we want to make m-1 the new end since m is *in* the interval,
i.e. (s, m-1]; as opposed to (s, m], because otherwise the currently
check, which is now known to not be the target, would remain in the
search interval.