[git] merge Delta with DeltaDecoder
Now that we have packReader we can have multiple open objects referring
to the same open pack file; this in turn allows us to have many 'open'
Delta objects, all with their own packReader. Therefore we can have the
decoding logic and most importantly the state in the Delta object.